Fix folder path
