+++ src/WinGetUtilInterop.UnitTests/TestCollateral/AllEqualityWithDescription.yaml
@@ -1,0 +1,28 @@
+Id: id.id
+Version: 1.2.3
+Publisher: test
+InstallerType: exe
+Description: This is a description
+Switches:
+  Custom: custom
+  Silent: /s
+  SilentWithProgress: /sp
+  Interactive: /i
+  Language: en-US
+  Log: /l=<LOGPATH>
+  InstallLocation: /d=<INSTALLPATH>
+Installers:
+  - Arch: x86
+    Sha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+    SignatureSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+    Language: en-US
+    InstallerType: msix
+    Scope: user
+    Switches:
+      Language: /en-US
+      Custom: /s
+  - Arch: x64
+    Url: https://contosa.net/publiccontainer/contosainstaller64.exe
+    Sha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C83
+    Language: en-US
+    Scope: user
