+++ src/WinGetUtilInterop.UnitTests/TestCollateral/SomeEqualityWithLocalization.yaml
@@ -1,0 +1,17 @@
+Id: id.id
+Version: 1.2.3
+Switches:
+  Custom: custom
+  Silent: /s
+Installers:
+  - Arch: x86
+    Sha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+    SignatureSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  - Arch: x64
+    Url: https://contosa.net/publiccontainer/contosainstaller64.exe
+    Sha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C83
+Localization:
+  - Language: es-MX
+    Description: Text to display for es-MX
+    Homepage: https://github.com/microsoft/msix-packaging/es-MX
+    LicenseUrl: https://github.com/microsoft/msix-packaging/blob/master/LICENSE-es-MX
