+++ src/WinGetUtilInterop.UnitTests/TestCollateral/V1ManifestMerged.yaml
@@ -1,0 +1,142 @@
+Author: Microsoft
+Capabilities:
+- internetClient
+Commands:
+- makemsix
+- makeappx
+Copyright: Copyright Microsoft Corporation
+CopyrightUrl: https://www.microsoft.com/msixsdk/copyright
+Dependencies:
+  ExternalDependencies:
+  - Outside dependencies
+  PackageDependencies:
+  - MinimumVersion: 1.0.0
+    PackageIdentifier: Microsoft.MsixSdkDep
+  WindowsFeatures:
+  - IIS
+  WindowsLibraries:
+  - VC Runtime
+Description: The MSIX SDK project is an effort to enable developers
+FileExtensions:
+- appx
+- msix
+- appxbundle
+- msixbundle
+InstallModes:
+- interactive
+- silent
+- silentWithProgress
+InstallerLocale: en-US
+InstallerSuccessCodes:
+- 1
+- 0x80070005
+InstallerSwitches:
+  Custom: /custom
+  InstallLocation: /dir=<INSTALLPATH>
+  Interactive: /interactive
+  Log: /log=<LOGPATH>
+  Silent: /silence
+  SilentWithProgress: /silentwithprogress
+  Upgrade: /upgrade
+InstallerType: zip
+Installers:
+- Architecture: x86
+  Capabilities:
+  - internetClientPreview
+  Commands:
+  - makemsixPreview
+  - makeappxPreview
+  Dependencies:
+    ExternalDependencies:
+    - Preview Outside dependencies
+    PackageDependencies:
+    - PackageIdentifier: Microsoft.MsixSdkDepPreview
+    WindowsFeatures:
+    - PreviewIIS
+    WindowsLibraries:
+    - Preview VC Runtime
+  FileExtensions:
+  - appxbundle
+  - msixbundle
+  - appx
+  - msix
+  InstallModes:
+  - interactive
+  InstallerLocale: en-GB
+  InstallerSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  InstallerSwitches:
+    Custom: /c
+    InstallLocation: /d=<INSTALLPATH>
+    Interactive: /i
+    Log: /l=<LOGPATH>
+    Silent: /s
+    SilentWithProgress: /sp
+    Upgrade: /u
+  InstallerType: msix
+  InstallerUrl: https://www.microsoft.com/msixsdk/msixsdkx86.msix
+  MinimumOSVersion: 10.0.1.0
+  PackageFamilyName: Microsoft.DesktopAppInstallerPreview_8wekyb3d8bbwe
+  Platform:
+  - Windows.Desktop
+  Protocols:
+  - protocol1preview
+  - protocol2preview
+  RestrictedCapabilities:
+  - runFullTrustPreview
+  Scope: user
+  SignatureSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  UpgradeBehavior: install
+- Architecture: x64
+  InstallerSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  InstallerType: exe
+  InstallerUrl: https://www.microsoft.com/msixsdk/msixsdkx64.exe
+  ProductCode: '{Bar}'
+License: MIT License
+LicenseUrl: https://www.microsoft.com/msixsdk/license
+Localization:
+- Author: Microsoft UK
+  Copyright: Copyright Microsoft Corporation UK
+  CopyrightUrl: https://www.microsoft.com/msixsdk/copyright/UK
+  Description: The MSIX SDK project is an effort to enable developers UK
+  License: MIT License UK
+  LicenseUrl: https://www.microsoft.com/msixsdk/license/UK
+  PackageLocale: en-GB
+  PackageName: MSIX SDK UK
+  PackageUrl: https://www.microsoft.com/msixsdk/home/UK
+  PrivacyUrl: https://www.microsoft.com/privacy/UK
+  Publisher: Microsoft UK
+  PublisherSupportUrl: https://www.microsoft.com/support/UK
+  PublisherUrl: https://www.microsoft.com/UK
+  ShortDescription: This is MSIX SDK UK
+  Tags:
+  - appxsdkUK
+  - msixsdkUK
+ManifestType: merged
+ManifestVersion: 1.0.0
+MinimumOSVersion: 10.0.0.0
+Moniker: msixsdk
+PackageFamilyName: Microsoft.DesktopAppInstaller_8wekyb3d8bbwe
+PackageIdentifier: microsoft.msixsdk
+PackageLocale: en-US
+PackageName: MSIX SDK
+PackageUrl: https://www.microsoft.com/msixsdk/home
+PackageVersion: 1.7.32
+Platform:
+- Windows.Desktop
+- Windows.Universal
+PrivacyUrl: https://www.microsoft.com/privacy
+ProductCode: '{Foo}'
+Protocols:
+- protocol1
+- protocol2
+Publisher: Microsoft
+PublisherSupportUrl: https://www.microsoft.com/support
+PublisherUrl: https://www.microsoft.com
+RestrictedCapabilities:
+- runFullTrust
+Scope: machine
+ShortDescription: This is MSIX SDK
+Tags:
+- appxsdk
+- msixsdk
+UpgradeBehavior: uninstallPrevious
