+++ src/WinGetUtilInterop.UnitTests/TestCollateral/V1_1ManifestMerged.yaml
@@ -1,0 +1,187 @@
+Agreements:
+- Agreement: DefaultText
+  AgreementLabel: DefaultLabel
+  AgreementUrl: https://DefaultAgreementUrl.net
+AppsAndFeaturesEntries:
+- DisplayName: DisplayName
+  DisplayVersion: DisplayVersion
+  InstallerType: exe
+  ProductCode: ProductCode
+  Publisher: Publisher
+  UpgradeCode: UpgradeCode
+Author: Microsoft
+Capabilities:
+- internetClient
+Commands:
+- makemsix
+- makeappx
+Copyright: Copyright Microsoft Corporation
+CopyrightUrl: https://www.microsoft.com/msixsdk/copyright
+Dependencies:
+  ExternalDependencies:
+  - Outside dependencies
+  PackageDependencies:
+  - MinimumVersion: 1.0.0
+    PackageIdentifier: Microsoft.MsixSdkDep
+  WindowsFeatures:
+  - IIS
+  WindowsLibraries:
+  - VC Runtime
+Description: The MSIX SDK project is an effort to enable developers
+ElevationRequirement: elevatesSelf
+ExpectedReturnCodes:
+- InstallerReturnCode: 10
+  ReturnResponse: packageInUse
+FileExtensions:
+- appx
+- msix
+- appxbundle
+- msixbundle
+InstallLocationRequired: true
+InstallModes:
+- interactive
+- silent
+- silentWithProgress
+InstallerAbortsTerminal: true
+InstallerLocale: en-US
+InstallerSuccessCodes:
+- 1
+- 0x80070005
+InstallerSwitches:
+  Custom: /custom
+  InstallLocation: /dir=<INSTALLPATH>
+  Interactive: /interactive
+  Log: /log=<LOGPATH>
+  Silent: /silence
+  SilentWithProgress: /silentwithprogress
+  Upgrade: /upgrade
+InstallerType: zip
+Installers:
+- Architecture: x86
+  Capabilities:
+  - internetClientPreview
+  Commands:
+  - makemsixPreview
+  - makeappxPreview
+  Dependencies:
+    ExternalDependencies:
+    - Preview Outside dependencies
+    PackageDependencies:
+    - PackageIdentifier: Microsoft.MsixSdkDepPreview
+    WindowsFeatures:
+    - PreviewIIS
+    WindowsLibraries:
+    - Preview VC Runtime
+  ElevationRequirement: elevationRequired
+  ExpectedReturnCodes:
+  - InstallerReturnCode: 2
+    ReturnResponse: contactSupport
+  FileExtensions:
+  - appxbundle
+  - msixbundle
+  - appx
+  - msix
+  InstallLocationRequired: false
+  InstallModes:
+  - interactive
+  InstallerAbortsTerminal: false
+  InstallerLocale: en-GB
+  InstallerSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  InstallerSwitches:
+    Custom: /c
+    InstallLocation: /d=<INSTALLPATH>
+    Interactive: /i
+    Log: /l=<LOGPATH>
+    Silent: /s
+    SilentWithProgress: /sp
+    Upgrade: /u
+  InstallerType: msix
+  InstallerUrl: https://www.microsoft.com/msixsdk/msixsdkx86.msix
+  Markets:
+    ExcludedMarkets:
+    - US
+  MinimumOSVersion: 10.0.1.0
+  PackageFamilyName: Microsoft.DesktopAppInstallerPreview_8wekyb3d8bbwe
+  Platform:
+  - Windows.Desktop
+  Protocols:
+  - protocol1preview
+  - protocol2preview
+  ReleaseDate: 2021-02-02
+  RequireExplicitUpgrade: false
+  RestrictedCapabilities:
+  - runFullTrustPreview
+  Scope: user
+  SignatureSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  UnsupportedOSArchitectures:
+  - arm64
+  UpgradeBehavior: install
+- Architecture: x64
+  InstallerSha256: 69D84CA8899800A5575CE31798293CD4FEBAB1D734A07C2E51E56A28E0DF8C82
+  InstallerType: exe
+  InstallerUrl: https://www.microsoft.com/msixsdk/msixsdkx64.exe
+  ProductCode: '{Bar}'
+License: MIT License
+LicenseUrl: https://www.microsoft.com/msixsdk/license
+Localization:
+- Agreements:
+  - Agreement: Text
+    AgreementLabel: Label
+    AgreementUrl: https://AgreementUrl.net
+  Author: Microsoft UK
+  Copyright: Copyright Microsoft Corporation UK
+  CopyrightUrl: https://www.microsoft.com/msixsdk/copyright/UK
+  Description: The MSIX SDK project is an effort to enable developers UK
+  License: MIT License UK
+  LicenseUrl: https://www.microsoft.com/msixsdk/license/UK
+  PackageLocale: en-GB
+  PackageName: MSIX SDK UK
+  PackageUrl: https://www.microsoft.com/msixsdk/home/UK
+  PrivacyUrl: https://www.microsoft.com/privacy/UK
+  Publisher: Microsoft UK
+  PublisherSupportUrl: https://www.microsoft.com/support/UK
+  PublisherUrl: https://www.microsoft.com/UK
+  ReleaseNotes: Release notes
+  ReleaseNotesUrl: https://ReleaseNotes.net
+  ShortDescription: This is MSIX SDK UK
+  Tags:
+  - appxsdkUK
+  - msixsdkUK
+ManifestType: merged
+ManifestVersion: 1.1.0
+Markets:
+  AllowedMarkets:
+  - US
+MinimumOSVersion: 10.0.0.0
+Moniker: msixsdk
+PackageFamilyName: Microsoft.DesktopAppInstaller_8wekyb3d8bbwe
+PackageIdentifier: microsoft.msixsdk
+PackageLocale: en-US
+PackageName: MSIX SDK
+PackageUrl: https://www.microsoft.com/msixsdk/home
+PackageVersion: 1.7.32
+Platform:
+- Windows.Desktop
+- Windows.Universal
+PrivacyUrl: https://www.microsoft.com/privacy
+ProductCode: '{Foo}'
+Protocols:
+- protocol1
+- protocol2
+Publisher: Microsoft
+PublisherSupportUrl: https://www.microsoft.com/support
+PublisherUrl: https://www.microsoft.com
+ReleaseDate: 2021-01-01
+ReleaseNotes: Default release notes
+ReleaseNotesUrl: https://DefaultReleaseNotes.net
+RequireExplicitUpgrade: true
+RestrictedCapabilities:
+- runFullTrust
+Scope: machine
+ShortDescription: This is MSIX SDK
+Tags:
+- appxsdk
+- msixsdk
+UnsupportedOSArchitectures:
+- arm
+UpgradeBehavior: uninstallPrevious
