--- src/WinGetUtilsInterop.UnitTests/TestCollateral/SomeEqualityWithoutInstallers.yaml
@@ -1,5 +1,0 @@
-Id: id.id
-Version: 1.2.3
-Switches:
-  Custom: custom
-  Silent: /s
